Add httPutResponseHopLimit for the EC2 instance
Instance metadata service limits the number of hops a request can make,
and with the AWS SDK & Dynamo update the dynamo client uses IMDsv2,
which limits the number of hops to 1. With this creating a dynamo db client
would time out, I assume there is an extra hop due to the Docker container.

Setting the httpPutResponseHopLimit to 2 resolves the issue.